Moving out BlobWriteChannel tests
